Move solvers to a new namespace
